requirements.txt
azure-cognitiveservices-vision-computervision
pillow
requests
matplotlib
flask
azure-storage-blob
opencv-python
azure-cognitiveservices-speech